refactor: centralize duplicated test fixtures in conftest.py
Move temp_registry_dir, temp_registry, and mock_args fixtures to
conftest.py. Remove duplicates from 6 test files where the implementation
matched. Files with specialized variants (yield+cleanup, custom args)
are left as-is.
