+++ tests/test_controller_registry_sync.py
@@ -6,21 +6,12 @@ from unittest.mock import patch
 import pytest
 
 from synapse.controller import TerminalController
-from synapse.registry import AgentRegistry
 
 
 @pytest.fixture
 def temp_registry_dir(tmp_path):
     """Create a temporary registry directory."""
     return tmp_path
-
-
-@pytest.fixture
-def temp_registry(temp_registry_dir):
-    """Create a test registry with temp directory."""
-    reg = AgentRegistry()
-    reg.registry_dir = temp_registry_dir
-    return reg
 
 
 class TestStatusSyncToRegistry:
